+++ app/layout.tsx
@@ -6,8 +6,8 @@ import "./globals.css";
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "UoA Degree Planner",
+  description: "UoA Degree Planner",
 };
 
 export const dynamic = 'force-dynamic';
