+++ app/page.tsx
@@ -1,5 +1,9 @@
 import Link from "next/link";
 export const dynamic = "force-dynamic";
+import checkIcon from "@/assets/checkIcon";
+import courseIcon from "@/assets/courseIcon";
+import contactIcon from "@/assets/contactIcon";
+import arrowIcon from "@/assets/arrowIcon";
 
 export default async function Index() {
   return (
@@ -23,18 +27,7 @@ export default async function Index() {
           <div className="flex flex-wrap lg:w-[580px] sm:mx-auto sm:mb-2 -mx-2 ">
             <div className="p-2 sm:w-1/2 w-full">
               <div className="bg-gray-100 rounded flex p-4 h-full items-center">
-                <svg
-                  fill="none"
-                  stroke="#3d74ff"
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth="3"
-                  className="text-indigo-500 w-6 h-6 flex-shrink-0 mr-4"
-                  viewBox="0 0 24 24"
-                >
-                  <path d="M22 11.08V12a10 10 0 11-5.93-9.14"></path>
-                  <path d="M22 4L12 14.01l-3-3"></path>
-                </svg>
+                {checkIcon()}
                 <span className="title-font font-medium text-lg">
                   {"Bachelor of Engineering (Honours)"}
                 </span>
@@ -42,25 +35,14 @@ export default async function Index() {
             </div>
             <div className="p-2 sm:w-1/2 w-full">
               <div className="bg-gray-100 rounded flex p-4 h-full items-center">
-                <svg
-                  fill="none"
-                  stroke="#3d74ff"
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth="3"
-                  className="text-indigo-500 w-6 h-6 flex-shrink-0 mr-4"
-                  viewBox="0 0 24 24"
-                >
-                  <path d="M22 11.08V12a10 10 0 11-5.93-9.14"></path>
-                  <path d="M22 4L12 14.01l-3-3"></path>
-                </svg>
+                {checkIcon()}
                 <span className="title-font font-medium text-lg">
                   Bachelor of Science
                 </span>
               </div>
             </div>
           </div>
-          <Link href="/planner">
+          <Link href="/pages/planner">
             <button className="flex mx-auto mt-16 text-white bg-[#3d74ff] border-0 py-3 px-10 focus:outline-none hover:bg-[#2853bf] rounded-full text-lg">
               Plan my degree
             </button>
@@ -74,17 +56,7 @@ export default async function Index() {
             <div className="p-4 lg:w-1/2 md:w-full">
               <div className="flex border-2 rounded-lg border-gray-200 border-opacity-50 p-8 sm:flex-row flex-col">
                 <div className="w-16 h-16 sm:mr-8 sm:mb-0 mb-4 inline-flex items-center justify-center rounded-full bg-indigo-100 text-indigo-500 flex-shrink-0">
-                  <svg
-                    fill="none"
-                    stroke="#3d74ff"
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth="2"
-                    className="w-8 h-8"
-                    viewBox="0 0 24 24"
-                  >
-                    <path d="M22 12h-4l-3 9L9 3l-3 9H2"></path>
-                  </svg>
+                  {courseIcon()}
                 </div>
                 <div className="flex-grow">
                   <h2 className="text-gray-900 text-lg title-font font-medium mb-3">
@@ -96,21 +68,11 @@ export default async function Index() {
                     and summaries on the "Courses" page.
                   </p>
                   <Link
-                    href="/courses"
+                    href="/pages/courses"
                     className="mt-3 text-[#3d74ff] inline-flex items-center text-2xl font-semibold"
                   >
                     Go to Courses
-                    <svg
-                      fill="none"
-                      stroke="#3d74ff"
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth="2"
-                      className="w-7 h-7 ml-2"
-                      viewBox="0 0 24 24"
-                    >
-                      <path d="M5 12h14M12 5l7 7-7 7"></path>
-                    </svg>
+                    {arrowIcon()}
                   </Link>
                 </div>
               </div>
@@ -118,18 +80,7 @@ export default async function Index() {
             <div className="p-4 lg:w-1/2 md:w-full">
               <div className="flex border-2 rounded-lg border-gray-200 border-opacity-50 p-8 sm:flex-row flex-col">
                 <div className="w-16 h-16 sm:mr-8 sm:mb-0 mb-4 inline-flex items-center justify-center rounded-full bg-indigo-100 text-indigo-500 flex-shrink-0">
-                  <svg
-                    fill="none"
-                    stroke="#3d74ff"
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth="2"
-                    className="w-10 h-10"
-                    viewBox="0 0 24 24"
-                  >
-                    <path d="M20 21v-2a4 4 0 00-4-4H8a4 4 0 00-4 4v2"></path>
-                    <circle cx="12" cy="7" r="4"></circle>
-                  </svg>
+                  {contactIcon()}
                 </div>
                 <div className="flex-grow">
                   <h2 className="text-gray-900 text-lg title-font font-medium mb-3">
@@ -140,21 +91,11 @@ export default async function Index() {
                     contact page for more information.
                   </p>
                   <Link
-                    href="/contact"
+                    href="/pages/contact"
                     className="mt-3 text-[#3d74ff] inline-flex items-center text-2xl font-semibold"
                   >
                     Learn More
-                    <svg
-                      fill="none"
-                      stroke="currentColor"
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth="2"
-                      className="w-7 h-78 ml-2"
-                      viewBox="0 0 24 24"
-                    >
-                      <path d="M5 12h14M12 5l7 7-7 7"></path>
-                    </svg>
+                    {arrowIcon()}
                   </Link>
                 </div>
               </div>
