+++ app/pages/contact/page.tsx
@@ -1,4 +1,4 @@
-import '../globals.css';
+import '../../globals.css';
 
 export default async function About() {
   return (
