+++ assets/arrowIcon.tsx
@@ -1,0 +1,15 @@
+export default function arrowIcon() {
+  return (
+    <svg
+      fill="none"
+      stroke="#3d74ff"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth="2"
+      className="w-7 h-7 ml-2"
+      viewBox="0 0 24 24"
+    >
+      <path d="M5 12h14M12 5l7 7-7 7"></path>
+    </svg>
+  );
+}
