+++ assets/contactIcon.tsx
@@ -1,0 +1,16 @@
+export default function contactIcon() {
+  return (
+    <svg
+      fill="none"
+      stroke="#3d74ff"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth="2"
+      className="w-10 h-10"
+      viewBox="0 0 24 24"
+    >
+      <path d="M20 21v-2a4 4 0 00-4-4H8a4 4 0 00-4 4v2"></path>
+      <circle cx="12" cy="7" r="4"></circle>
+    </svg>
+  );
+}
