+++ assets/courseIcon.tsx
@@ -1,0 +1,15 @@
+export default function courseIcon() {
+  return (
+    <svg
+      fill="none"
+      stroke="#3d74ff"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth="2"
+      className="w-8 h-8"
+      viewBox="0 0 24 24"
+    >
+      <path d="M22 12h-4l-3 9L9 3l-3 9H2"></path>
+    </svg>
+  );
+}
